# rope-manager-backend/openapi.yaml
openapi: 3.0.3
info:
  title: 绳包管理器后端 API
  description: |
    绳包管理器后端服务的完整API文档
    
    ## 认证方式
    系统支持两种认证方式，按优先级顺序：
    1. **Authorization 头部认证（推荐）**: `Authorization: Bearer <token>`
    2. **Cookie 认证**: `Cookie: auth_token=<token>`
    
    ## 权限等级
    - **普通用户 (user)**: 基础功能使用权限
    - **元老 (elder)**: 拥有部分管理权限，可审核资源
    - **版主 (moderator)**: 拥有内容管理权限
    - **管理员 (admin)**: 拥有所有系统权限
    
    ## API 响应格式
    所有API统一返回格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
    - code: 状态码，0表示成功，其他表示错误
    - message: 响应消息
    - data: 响应数据（可选）
    
  version: 1.0.0
  contact:
    name: 绳包管理器API支持
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:15201/api/v1
    description: 开发服务器 - API端点
  - url: http://127.0.0.1:15201
    description: 开发服务器 - 健康检查端点
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
        message:
          type: string
          description: 响应消息
        data:
          description: 响应数据（可选）
      required:
        - code
        - message

    # 用户相关模型
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        password_hash:
          type: string
          writeOnly: true
        nickname:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, moderator, elder, user]
        star:
          type: integer
          description: 用户星级
        ban_status:
          type: string
          enum: [normal, suspended, banned]
        ban_reason:
          type: string
          nullable: true
        qq_number:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        login_count:
          type: integer
        upload_count:
          type: integer
        download_count:
          type: integer
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
        is_admin:
          type: boolean
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        nickname:
          type: string
          nullable: true
        qq_number:
          type: string
          nullable: true
        verification_code:
          type: string
          description: 邮箱验证码
      required:
        - username
        - email
        - password
        - verification_code
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        nickname:
          type: string
        star:
          type: integer
        ban_status:
          type: string
          enum: [normal, suspended, banned]
        ban_reason:
          type: string
        role:
          type: string
          enum: [admin, moderator, elder, user]
        qq_number:
          type: string
        avatar_url:
          type: string

    # 绳包相关模型
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        author:
          type: string
        version:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        file_url:
          type: string
        file_size:
          type: integer
          format: int64
          nullable: true
        download_count:
          type: integer
        like_count:
          type: integer
        favorite_count:
          type: integer
        category_id:
          type: integer
          nullable: true
        status:
          type: string
          enum: [Pending, Active, Rejected, Inactive, Deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        reviewer_id:
          type: integer
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        review_comment:
          type: string
          nullable: true
        is_pinned:
          type: boolean
        is_featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
          nullable: true
...
    PackageListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 评论相关模型
    Comment:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        target_type:
          type: string
          enum: [Package, User, System]
        target_id:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [Active, Hidden, Deleted]
        parent_id:
          type: integer
          nullable: true
        likes:
          type: integer
        dislikes:
          type: integer
        pinned:
          type: boolean
        author_name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        author_role:
          type: string
          nullable: true
        author_avatar:
          type: string
          nullable: true
        author_qq:
          type: string
          nullable: true
        target_title:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CommentListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 帖子相关模型
...
    BatchIdsRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
      required:
        - ids
...
paths:
...
  /users:
    get:
      tags:
        - 用户管理
      summary: 获取用户列表
      description: 获取所有用户列表（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          total:
                            type: integer
                          page:
                            type: integer
                          pageSize:
                            type: integer
                          totalPages:
                            type: integer
        '403':
          description: 权限不足

    post:
      tags:
        - 用户管理
      summary: 创建用户
      description: 创建新用户（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: 创建成功
        '403':
          description: 权限不足
  /users/batch:
    delete:
      tags:
        - 用户管理
      summary: 批量删除用户
      description: 批量删除用户（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchIdsRequest'
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
  /users/profile:
    get:
      tags:
        - 用户管理
      summary: 获取当前用户资料
      description: 获取当前登录用户的个人资料
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: 未授权

    put:
      tags:
        - 用户管理
      summary: 更新当前用户资料
      description: 更新当前登录用户的个人资料
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
        '401':
          description: 未授权
  /users/my-resources:
    get:
      tags:
        - 用户管理
      summary: 获取我的资源
      description: 获取当前用户上传的所有资源
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PackageListResponse'
        '401':
          description: 未授权
  /users/my-comments:
    get:
      tags:
        - 用户管理
      summary: 获取我的评论
      description: 获取当前用户的所有评论
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'
        '401':
          description: 未授权
  /users/change-password:
    post:
      tags:
        - 用户管理
      summary: 修改密码
      description: 修改当前用户密码
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
                  minLength: 6
              required:
                - old_password
                - new_password
      responses:
        '200':
          description: 修改成功
        '400':
          description: 修改失败
        '401':
          description: 未授权
  /users/{id}:
    get:
      tags:
        - 用户管理
      summary: 获取用户信息
      description: 根据ID获取用户信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          description: 用户不存在

    put:
      tags:
        - 用户管理
      summary: 更新用户信息
      description: 更新指定用户信息（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 权限不足
        '404':
          description: 用户不存在

    delete:
      tags:
        - 用户管理
      summary: 删除用户
      description: 删除指定用户（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
        '404':
          description: 用户不存在
  /users/{id}/comments:
    get:
      tags:
        - 用户管理
      summary: 获取用户评论
      description: 获取指定用户的所有评论
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'

  # 绳包管理接口
...